# docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        error:
          example: validation failed
          type: string
      type: object
...
    attacker.Attack:
      properties:
        config:
          $ref: '#/components/schemas/attacker.AttackConfig'
        result:
          $ref: '#/components/schemas/attacker.AttackResult'
        started_at:
          type: string
        status:
          type: string
      type: object
    attacker.AttackConfig:
      properties:
        dedup_bypass:
          type: string
        duration:
          $ref: '#/components/schemas/time.Duration'
        experiment_id:
          type: string
        id:
          type: string
        meta_trace_id:
          type: string
        rate:
          type: integer
        run_ref:
          type: string
        target:
          $ref: '#/components/schemas/attacker.TargetSpec'
        workflow_label:
          type: string
      type: object
    attacker.AttackResult:
      properties:
        bytes_in:
          $ref: '#/components/schemas/attacker.BytesSummary'
        bytes_out:
          $ref: '#/components/schemas/attacker.BytesSummary'
        duration:
          $ref: '#/components/schemas/time.Duration'
        errors:
          items:
            type: string
          type: array
          uniqueItems: false
        latencies:
          $ref: '#/components/schemas/attacker.LatencySummary'
        rate_actual:
          type: number
        status_codes:
          additionalProperties:
            type: integer
          type: object
        success:
          type: number
        total_requests:
          type: integer
      type: object
    attacker.BytesSummary:
      properties:
        mean:
          type: number
        total:
          type: integer
      type: object
    attacker.LatencySummary:
      properties:
        max:
          $ref: '#/components/schemas/time.Duration'
        min:
          $ref: '#/components/schemas/time.Duration'
        p50:
          $ref: '#/components/schemas/time.Duration'
        p90:
          $ref: '#/components/schemas/time.Duration'
        p95:
          $ref: '#/components/schemas/time.Duration'
        p99:
          $ref: '#/components/schemas/time.Duration'
      type: object
    attacker.TargetSpec:
      properties:
        body:
          items:
            type: integer
          type: array
          uniqueItems: false
        headers:
          additionalProperties:
            type: string
          type: object
        method:
          type: string
        url:
          type: string
      type: object
...
    time.Duration:
      enum:
      - -9223372036854775808
      - 9223372036854775807
      - 1
      - 1000
      - 1000000
      - 1000000000
      - 60000000000
      - 3600000000000
      type: integer
      x-enum-varnames:
      - minDuration
      - maxDuration
      - Nanosecond
      - Microsecond
      - Millisecond
      - Second
      - Minute
      - Hour
...
info:
  description: 'Zeus is the execution plane for faults-lab. Manteion issues

    commands to it. This document is the integration contract between

    the two. Run lifecycle state machine and cross-cutting concerns

    (dataset handshake, error envelope) are documented in

    docs/api-contract.md.'
  title: Zeus Load Generation API
  version: '1.0'
openapi: 3.1.0
paths:
  /attacks:
    get:
      description: 'Returns a bare array of attacks. Optional query filters narrow
        by status,

        experiment_id, and run_ref.'
      parameters:
      - description: filter by attack status (e.g. running, completed)
        in: query
        name: status
        schema:
          type: string
      - description: filter by experiment id
        in: query
        name: experiment_id
        schema:
          type: string
      - description: filter by linked run id
        in: query
        name: run_ref
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/attacker.Attack'
                type: array
          description: OK
      summary: List attacks
      tags:
      - attacks
    post:
      description: 'Persist and launch a precision attack. The body is an attacker.AttackConfig.

        When run_ref references an existing run, meta_trace_id and workflow_label
        are

        inherited from the linked run if not set on the request. The response is an

        inline envelope with id, status, and started_at; the launched Attack object
        is

        retrievable via GET /attacks/{id}.'
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/attacker.AttackConfig'
                description: attack configuration
                summary: attack
        description: attack configuration
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: attack accepted; envelope contains id, status, started_at
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: invalid JSON or attack validation error
      summary: Launch attack
      tags:
      - attacks
  /attacks/{id}:
    delete:
      description: Cancels a running attack and decrements the active-attacks gauge.
      parameters:
      - description: Attack ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: attack stopped
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: attack not found
      summary: Stop attack
      tags:
      - attacks
    get:
      parameters:
      - description: Attack ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/attacker.Attack'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: attack not found
      summary: Get attack
      tags:
      - attacks
  /attacks/{id}/stats:
    get:
      description: 'Returns an inline envelope with attack_id, status, config, optional
        result

        (vegeta-derived totals + latency percentiles), and optional dedup strategy

        metadata when DedupBypass is configured on the attack.'
      parameters:
      - description: Attack ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: attack stats envelope (attack_id, status, config, result?,
            dedup?)
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: attack not found
      summary: Get attack stats
      tags:
      - attacks
...
servers:
- description: Local dev (HTTP)
  url: http://localhost:8080/api/v1
- description: Local dev (HTTPS)
  url: https://localhost:8080/api/v1